requirements.txt
Flask==1.1.2
flask-cors==3.0.8
pytest==5.4.1
testfixtures==6.14.1
watchdog~=2.0.2
pypubsub~=4.0.3
busypie~=0.4.5
ddt~=1.4.2